+++ src/AppConstants.ts
@@ -1,9 +1,0 @@
-export default class AppConstants {
-    public static readonly ACCEPTED_FILENAMES = [
-        'todo.txt',
-        'done.txt'
-    ]
-
-    public static ARCHIVE_FILENAME = 'done.txt';
-    public static TODO_FILENAME = 'todo.txt';
-}